Glove
The aim of the project is to make a gesture-controlled crane. The crane will be
controlled by particular movements of the hand, and these will be detected by an
accelerometer and sent to the crane using RF signals and the crane will move
accordingly. It can grip and carry objects of suitable width it with it and drop it
when asked to do so. There will be a glove worn on the hand containing
components of accelerometer and flex sensor, microcontroller, and a wireless
transmitter module. The accelerometer and flex sensor will detect the accelerations
along its axes and this will be transmitted to the transmitter through a
microcontroller . The other wireless module, on the crane, that is, the transceiver
will receive these signals and send it to the microcontroller, which will be coded to
the crane as desired. The hand gestures could also instruct the crane to grip objects
and take it off with it. The gripping mechanism is controlled by a servo motor. The
lifting mechanism is controlled by five high torque servo motors.